+++ client/js/lib/beseda/transport/request/XHRRequest.js
@@ -2,12 +2,29 @@
  * @constructor
  * @extends BesedaPackage.events.EventEmitter
  */
-BesedaPackage.transport.request.XHRRequest = function(method) {
+BesedaPackage.transport.request.XHRRequest = function(method, xDomain) {
     BesedaPackage.events.EventEmitter.prototype.constructor.call(this);
+
+    this._isXDomain = xDomain;
 
     this.url = null;
     this.method = method;
     this.data = null;
+};
+
+BesedaPackage.transport.request.XHRRequest.createRequest = function(xDomain) {
+    var request = null;
+
+    if (window.XMLHttpRequest && (!xDomain || (new XMLHttpRequest()).withCredentials != undefined)) {
+        request = new XMLHttpRequest();
+    } else if (window.XDomainRequest && xDomain) {
+        request = new XDomainRequest();
+        request.onprocess = function(){};
+    } else if (window.ActiveXObject && !xDomain) {
+        request = new ActiveXObject('Microsoft.XMLHTTP');
+    }
+
+    return request;
 };
 
 BesedaPackage.utils.inherits(BesedaPackage.transport.request.XHRRequest, BesedaPackage.events.EventEmitter);
@@ -22,8 +39,8 @@ BesedaPackage.transport.request.XHRRequest.prototype.send = function(url) {
     }
 
     var requestURL = this.url + '/' + (new Date().getTime());
-    var request = this.__createRequest();
-	
+    var request = BesedaPackage.transport.request.XHRRequest.createRequest(this._isXDomain);
+
     var self = this;
     request.onreadystatechange = function() {
         self.__requestStateHandler(request);
@@ -57,21 +74,4 @@ BesedaPackage.transport.request.XHRRequest.prototype.__requestStateHandler = fun
         request.abort();
 	    request = null;
     }
-};
-
-BesedaPackage.transport.request.XHRRequest.prototype.__createRequest = function() {
-	var request =  null;
-
-	if (!+'\v1') {
-		if (window.XDomainRequest) {
-			request = new XDomainRequest();
-			request.onprocess = function(){};
-		} else {
-			request = new ActiveXObject("Microsoft.XMLHTTP");
-		}
-	} else {
-		request = new XMLHttpRequest();
-	}
-
-	return request;
 };